Add files via upload
House listing down to 3 houses, 3 images added images folder in the path.
Index2 contains buttons and an interaction list with script.js. Put the dates on the same side as the buttons in style.css.
We have to merge with Isak's version for the JS. Works only when opening localhost server for the Json data to be fetched.